requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.3
django-browser-reload==1.17.0
django-environ==0.11.2
psycopg2==2.9.10
pycparser==2.22
python-dotenv==1.0.1
sqlparse==0.5.2
